+++ PersistencePlugin/src/main/java/com/elmakers/mine/bukkit/persistence/dao/MigrationStep.java
@@ -52,7 +52,7 @@ public class MigrationStep extends Persisted
 		
 		
 		entity = entityClass.getName();
-		schema = entityClass.getSchema();
+		schema = entityClass.getSchemaName();
 	}
 	
 	/**