+++ PersistencePlugin/src/main/java/com/elmakers/mine/craftbukkit/persistence/core/Schema.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
+import com.elmakers.mine.craftbukkit.persistence.data.DataStore;
+
 /**
  * Describes a schema.
  * 
@@ -14,6 +16,12 @@ import java.util.List;
  */
 public class Schema
 {	
+	public Schema(String name, DataStore defaultStore)
+	{
+		this.name = name;
+		this.defaultStore = defaultStore;
+	}
+	
 	public String getName()
 	{
 		return name;
@@ -40,7 +48,21 @@ public class Schema
 		return nameMap.get(className);
 	}
 	
-	private String name;
-	private final List<PersistedClass> persistedClasses = new ArrayList<PersistedClass>();
-	private final HashMap<String, PersistedClass> nameMap = new HashMap<String, PersistedClass>();
+	public DataStore getStore()
+	{
+		return defaultStore;
+	}
+	
+	public void disconnect()
+	{
+		if (defaultStore != null)
+		{
+			defaultStore.disconnect();
+		}
+	}
+
+	private String									name;
+	private DataStore								defaultStore;
+	private final List<PersistedClass>				persistedClasses	= new ArrayList<PersistedClass>();
+	private final HashMap<String, PersistedClass>	nameMap				= new HashMap<String, PersistedClass>();
 }