+++ PersistencePlugin/src/main/java/com/elmakers/mine/craftbukkit/persistence/data/sql/SqlDataRow.java
@@ -29,7 +29,7 @@ public class SqlDataRow extends DataRow
 			{
 				int sqlType = rowInfo.getColumnType(i);
 				String columnName = rowInfo.getColumnName(i);
-				DataType dataType = getTypeFromSqlType(sqlType);
+				DataType dataType = DataType.getTypeFromSqlType(sqlType);
 				SqlDataField field = new SqlDataField(row, i, columnName, dataType);
 				add(field);
 			}
@@ -39,99 +39,5 @@ public class SqlDataRow extends DataRow
 			log.warning("Persistence: error creating SQLDataRow: " + ex.getMessage());
 		}
 		
-	}
-	
-	/**
-	 * Convert a SQL data type to a DataType.
-	 * 
-	 * Only converts supported DataTypes.
-	 * 
-	 * @param sqlType The java.sql.Type value to convert
-	 * @return A DataType value, or DataType.NULL if invalid
-	 */
-	public static DataType getTypeFromSqlType(int sqlType)
-	{
-		DataType dataType = DataType.NULL;
-		
-		switch(sqlType)
-		{
-			case java.sql.Types.BOOLEAN:
-				dataType = DataType.BOOLEAN;
-				break;
-			case java.sql.Types.BIT:
-				dataType = DataType.BOOLEAN;
-				break;	
-			case java.sql.Types.DATE:
-				dataType = DataType.DATE;
-				break;
-			case java.sql.Types.DECIMAL:
-				dataType = DataType.DOUBLE;
-				break;	
-			case java.sql.Types.DOUBLE:
-				dataType = DataType.DOUBLE;
-				break;	
-			case java.sql.Types.FLOAT:
-				dataType = DataType.DOUBLE;
-				break;	
-			case java.sql.Types.INTEGER:
-				dataType = DataType.INTEGER;
-				break;	
-			case java.sql.Types.JAVA_OBJECT:
-				dataType = DataType.OBJECT;
-				break;	
-			case java.sql.Types.LONGNVARCHAR:
-				dataType = DataType.STRING;
-				break;	
-			case java.sql.Types.LONGVARCHAR:
-				dataType = DataType.STRING;
-				break;	
-			case java.sql.Types.NCHAR:
-				dataType = DataType.STRING;
-				break;	
-			case java.sql.Types.NULL:
-				dataType = DataType.NULL;
-				break;	
-			case java.sql.Types.NUMERIC:
-				dataType = DataType.INTEGER;
-				break;	
-			case java.sql.Types.REAL:
-				dataType = DataType.DOUBLE;
-				break;	
-			case java.sql.Types.ROWID:
-				dataType = DataType.INTEGER;
-				break;	
-			case java.sql.Types.SMALLINT:
-				dataType = DataType.INTEGER;
-				break;	
-			case java.sql.Types.TIME:
-				dataType = DataType.DATE;
-				break;	
-			case java.sql.Types.TIMESTAMP:
-				dataType = DataType.DATE;
-				break;	
-			case java.sql.Types.TINYINT:
-				dataType = DataType.INTEGER;
-				break;	
-			case java.sql.Types.VARCHAR:
-				dataType = DataType.STRING;
-				break;	
-		}
-		
-		return dataType;
-	}
-	
-	public static int getSqlType(DataType dataType)
-	{
-		switch(dataType)
-		{
-			case BOOLEAN: return java.sql.Types.BOOLEAN;
-			case DATE: return java.sql.Types.DATE;
-			case DOUBLE: return java.sql.Types.DOUBLE;
-			case FLOAT: return java.sql.Types.FLOAT;
-			case INTEGER: return java.sql.Types.INTEGER;
-			case STRING: return java.sql.Types.VARCHAR;
-		}
-		
-		return java.sql.Types.NULL;
 	}
 }